Fix media server link
