+++ frontend/app/globals.css
@@ -30,7 +30,7 @@
 
 @layer base {
   :root {
-    --background: 0 0% 100%;
+    --background: 240 4.8% 95.9%;
     --foreground: 0 0% 3.9%;
     --card: 0 0% 100%;
     --card-foreground: 0 0% 3.9%;
